Unroll xtea

+++ xtea/src/consts.rs
@@ -1,2 +1,2 @@
 pub const DELTA: u32 = 0x9e3779b9;
-pub const ROUNDS: u32 = 32;
+pub const ROUNDS: usize = 32;
